+++ docs/CHANGELOG.rst
@@ -14,6 +14,8 @@ Added
 -----
 - Add new entity annotation framework
 - Add set_permission method on PermissionQuerySet
+- Add notify_create method on Subscription model in observers to enable
+  sending notification when object is created
 
 
 ===================